--- app/jobs/relance_utilisateur_pour_campagne_job.rb
@@ -1,11 +1,0 @@
-# frozen_string_literal: true
-
-class RelanceUtilisateurPourCampagneJob < ApplicationJob
-  queue_as :default
-
-  def perform(campagne:)
-    return unless campagne.nombre_evaluations.zero?
-
-    CampagneMailer.with(campagne: campagne).relance.deliver_now
-  end
-end
