+++ app/views/compte_mailer/relance.html.erb
@@ -1,6 +1,7 @@
 <%= md t('.intro', prenom: @compte.prenom,
+                   delai_relance: distance_of_time_in_words(Compte::DELAI_RELANCE_NON_ACTIVATION),
                    cible_evaluation: @cible_evaluation) %>
-<%= md t('.x_personnes_utilisent_eva', count: @nombre_collegue) %>
+<%= md t('.x_personnes_utilisent_eva', count: @effectif) %>
 <%= md t('.besoin_aide', cible_evaluation: @cible_evaluation,
                          lien_rubrique_aide: admin_aide_url,
                          lien_livestorm: Eva::LIEN_LIVESTORM) %>
