--- config/locales/views/campagne_mailer.yml
@@ -1,19 +1,0 @@
-fr:
-  campagne_mailer:
-    relance:
-      objet: "%{prenom}, quelques conseils pour démarrer sur eva - Positionnement des compétences"
-      intro: |
-        ### Bonjour %{prenom},
-
-        Il y a 30 jours vous avez créé un compte sur le service numérique eva. Vous vouliez certainement accompagner différemment vos bénéficiaires. Nous vous rappelons qu'eva permet de détecter l'illettrisme et de valoriser vos "%{cible_evaluation}" au travers des compétences transversales acquises.
-      x_personnes_utilisent_eva:
-        zero: ""
-        one: "Déjà 1 collègue de votre structure ont créé leur compte."
-        other: "Déjà %{count} collègues de votre structure ont créé leur compte."
-      besoin_aide: |
-        ## Besoin d'aide ?
-
-        Pour accompagner les %{cible_evaluation} vers l’insertion professionnelle et vous servir d’eva, contacter notre équipe :
-        [support@eva.beta.gouv.fr](mailto:support@eva.beta.gouv.fr)
-
-        Vous pouvez aussi vous réinscrire à [une de nos présentations](%{lien_livestorm}) ou bien retrouver toute les informations dans la [rubrique aide](%{lien_rubrique_aide}).
