+++ spec/mailers/compte_mailer_spec.rb
@@ -27,4 +27,19 @@ describe CompteMailer, type: :mailer do
       expect(email.body).to include('paris2019')
     end
   end
+
+  describe 'relance' do
+    let(:structure) { Structure.new nom: 'Ma Super Structure' }
+    let(:compte) { Compte.new prenom: 'Paule', email: 'debut@test.com', structure: structure }
+    let(:campagne) { Campagne.new compte: compte }
+    let(:mail) { CompteMailer.with(compte: compte).relance }
+
+    it 'renders the headers' do
+      expect(mail.subject).to eq(
+        'Paule, quelques conseils pour démarrer sur eva - Positionnement des compétences'
+      )
+      expect(mail.to).to eq(['debut@test.com'])
+      expect(mail.from).to eql([Eva::EMAIL_CONTACT])
+    end
+  end
 end
